requirements/super_resolution.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0
tqdm==4.66.1
madgrad==1.3
matplotlib==3.8.1
timm==0.9.10
torchsr==1.0.4
clearml==1.13.2
pandas==2.1.3
pytest==7.4.3
natsort==8.4.0
colorama==0.4.6
python-box==7.1.1
scikit-learn==1.3.2
albumentations==1.3.1
pytorch-lightning==1.8.4
beautifulsoup4==4.12.2
transformers==4.36.2
lion-pytorch==0.1.2
streamlit==1.30.0
python-multipart==0.0.7
empatches==0.2.3
pillow==10.3.0
imagecodecs==2024.6.1
numpy==1.26.4